% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compute.R
\name{compute}
\alias{compute}
\title{Compute a new variable}
\usage{
compute(.data = NULL, .name = NULL, .var, .keep = "all")
}
\arguments{
\item{.data}{A tibble.}

\item{.name}{Nombre de la variable que vamos a crear.}
}
\value{
Devuelve un tibble.
}
\description{
Compute a new variable
}
\keyword{utils}
