% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correspondencia}
\alias{correspondencia}
\title{Correspondencia de CIIU revisión 3 a revisión 4}
\format{
Correspondencia de CIIU revisión 3 a revisión 4
}
\usage{
correspondencia
}
\description{
Correspondencia de CIIU revisión 3 a revisión 4
}
\keyword{ciiu}
