% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cotizaciones_ine_anual}
\alias{cotizaciones_ine_anual}
\title{Cotinzaciones del INE anuales}
\format{
Recetas de la ech
}
\usage{
cotizaciones_ine_anual
}
\description{
Cotinzaciones del INE anuales
}
\keyword{data_aux}
