% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cotizaciones_ine_mensual}
\alias{cotizaciones_ine_mensual}
\title{Cotinzaciones del INE mensuales}
\format{
Recetas de la ech
}
\usage{
cotizaciones_ine_mensual
}
\description{
Cotinzaciones del INE mensuales
}
\keyword{data_aux}
