% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dummy_cols.R
\name{dummy_cols}
\alias{dummy_cols}
\title{Fast creation of dummy variables}
\usage{
dummy_cols(
  .data,
  select_columns = NULL,
  remove_first_dummy = FALSE,
  remove_most_frequent_dummy = FALSE,
  ignore_na = FALSE,
  split = NULL,
  remove_selected_columns = FALSE,
  omit_colname_prefix = FALSE
)
}
\description{
Fast creation of dummy variables
}
