% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ech_recipes}
\alias{ech_recipes}
\alias{recipes}
\title{Recetas de la ECH}
\format{
Correspondencia de CIIU revisión 3 a revisión 4
}
\usage{
recipes
}
\description{
Recetas de la ECH
}
\keyword{ech}
\keyword{recipes}
