% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_employment_ai.R
\name{estimate_employment_ai.}
\alias{estimate_employment_ai.}
\title{Estimate the number of people engaged in innovation activities}
\usage{
estimate_employment_ai.(.data, .eaii, .by = NULL, .recipe = "ocupados_ai")
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.recipe}{Recipe name}

\item{.zap_labels}{Boolean of labels to be removed}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the number of people engaged in innovation activities
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_employment_ai.(
    .eaii = "2016-2018"
  )

}
\keyword{eaii}
\keyword{easy}
