% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_gap_gender.R
\name{estimate_gap_gender.}
\alias{estimate_gap_gender.}
\title{Estimate the number of people engaged in I+D activities}
\usage{
estimate_gap_gender.(.data, .eaii, .by = NULL, .recipe = "gap_gender")
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.by}{a variable to agroup the data}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the number of people engaged in I+D activities
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_professionals_id.(
    .eaii = "2016-2018"
  )

load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_professionals_id.(
    .eaii = "2016-2018",
    .recipe = "profesionales_nivel"
  )
}
\keyword{eaii}
\keyword{easy}
