% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_investment_ai.R
\name{estimate_investment_ai.}
\alias{estimate_investment_ai.}
\title{Estimate investment in innovation activities for the period}
\usage{
estimate_investment_ai.(
  .data,
  .eaii,
  .by = NULL,
  .recipe = "inversion",
  type = "anii",
  summarise = FALSE
)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.by}{Variable to group by}

\item{.recipe}{Recipe name}

\item{type}{Boolean of labels to be removed}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate investment in innovation activities for the period
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_investment_ai.(
    .eaii = "2016-2018",
    .by = "sector",
    type = "anii"
  )

}
\keyword{eaii}
\keyword{easy}
