% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-estimate_mean.R
\name{estimate_mean}
\alias{estimate_mean}
\alias{stats_mean.}
\title{Estimate a mean from a sample of data with H-T}
\usage{
stats_mean.(.data, .vars, .by = NULL)
}
\arguments{
\item{.by}{a variable to agroup the data}

\item{.weight}{weight}
}
\description{
Estimate a mean from a sample of data with H-T
}
\keyword{stats}
