% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-estimate_mean.R
\name{estimate_proportion}
\alias{estimate_proportion}
\alias{stats_proportion.}
\title{Estimate a mean from a sample of data with H-T}
\usage{
stats_proportion.(.data, .variable = NULL, .by = NULL, .filter_labels = NULL)
}
\arguments{
\item{.variable}{variable}

\item{.by}{a variable to agroup the data}

\item{.weight}{weight}
}
\description{
Estimate a mean from a sample of data with H-T
}
\keyword{stats}
