% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_ratio_innovative.R
\name{estimate_proportion_innovative.}
\alias{estimate_proportion_innovative.}
\title{Estimate the percentage of innovative companies}
\usage{
estimate_proportion_innovative.(
  .data,
  .eaii,
  .by = NULL,
  .recipe = "innovative_bussiness"
)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.by}{a variable to agroup the data}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the percentage of innovative companies
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_ratio_innovative.(
    .eaii = "2016-2018"
  )
}
\keyword{eaii}
\keyword{easy}
