% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_ratio_activities.R
\name{estimate_ratio_activities.}
\alias{estimate_ratio_activities.}
\title{Estimate the proportion of innovative companies by type of activity}
\usage{
estimate_ratio_activities.(
  .data,
  .eaii,
  .by = NULL,
  .recipe = "actividades",
  .vars = c("IDinterna", "IDexterna", "bienescapital", "hardware", "software",
    "transferencia", "propiedadintelectual", "ingenieria", "gestion", "capacitacion",
    "marketing", "TICS", "id", "ad_software")
)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.by}{a variable to agroup the data}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the proportion of innovative companies by type of activity
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_ratio_activities.(
    eaii = "2016-2018"
  )
}
\keyword{eaii}
\keyword{easy}
