% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_ratio_innovation.R
\name{estimate_ratio_innovation.}
\alias{estimate_ratio_innovation.}
\title{Estimate the proportion of the type of innovation}
\usage{
estimate_ratio_innovation.(
  .data,
  .eaii,
  .vars = c("producto", "proceso", "organizacion", "comercializacion"),
  .by = NULL,
  .recipe = "tipo_innovacion"
)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.vars}{Variables to calculate}

\item{.by}{a variable to agroup the data}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the proportion of the type of innovation
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_ratio_innovation.(
    .eaii = "2016-2018"
  )
}
\keyword{eaii}
\keyword{easy}
