% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii-obstacles_to_innovation.R
\name{estimate_ratio_obstacle_innovation.}
\alias{estimate_ratio_obstacle_innovation.}
\title{Estimate the ratio obstacle to innovation}
\usage{
estimate_ratio_obstacle_innovation.(
  .data,
  .eaii,
  .by = "realiza_innovacion",
  .recipe = "obstaculos",
  .vars = c("personal_capacitado", "rigidez", "riesgos", "retorno", "tamanio_mercado",
    "oportunidades_tecno", "financiamiento", "cooperacion", "insuficiencia_mercados",
    "insuficiencia_tecno", "desarrollo_instituciones", "infraestructura",
    "prop_deficiente", "inestabilidad_macro", "otros")
)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.by}{a variable to agroup the data}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the ratio obstacle to innovation
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_ratio_obstacle_innovation.(
    .eaii = "2016-2018"
  )
}
\keyword{eaii}
\keyword{easy}
