% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_public_support.R
\name{estimate_ratio_public_support.}
\alias{estimate_ratio_public_support.}
\title{Estimate the proportion of public support}
\usage{
estimate_ratio_public_support.(
  .data,
  .eaii,
  .by = NULL,
  .recipe = "apoyopublico"
)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.by}{a variable to agroup the data}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the proportion of public support
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_ratio_public_support.(
    .eaii = "2016-2018"
  )
}
\keyword{eaii}
\keyword{easy}
