% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-eaii_result_innovative.R
\name{estimate_ratio_result_innovation.}
\alias{estimate_ratio_result_innovation.}
\title{Estimate the result of innovation}
\usage{
estimate_ratio_result_innovation.(
  .data,
  .eaii,
  .by = NULL,
  .recipe = "resultado"
)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.eaii}{Survey year}

\item{.by}{a variable to agroup the data}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Estimate the result of innovation
}
\examples{


devtools::load_all()
load_base.(
  .eaii = "2016-2018"
) \%>\%
  estimate_ratio_result_innovation.(
    .eaii = "2016-2018"
  )
}
\keyword{eaii}
\keyword{easy}
