% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-to_ciiu4.R
\name{get_ciiu4.}
\alias{get_ciiu4.}
\title{Clasificación de CIIU Revisión 3 a CIIU Revisión 4 adaptada por el INE.}
\usage{
get_ciiu4.(.data, .ciiu3, .ID)
}
\arguments{
\item{.data}{A tibble.}

\item{.ciiu3}{Variable que tenga la clasificación ciiu3.}

\item{.ID}{número identificatorio del individuo de la muestra.}
}
\value{
Devuelve un tibble.
}
\description{
Clasificación de CIIU Revisión 3 a CIIU Revisión 4 adaptada por el INE.
}
\examples{
load_base.(
  "2007-2009-ser"
) \%>\%
  to_ciiu4.(
    ClaseAct,
    "INE"
  )
}
\keyword{utils}
