% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-load_base.R
\name{load_base.}
\alias{load_base.}
\title{Crear un grupo para luego particionar la poblaci'on y calcular el total poblacional en cada momento}
\usage{
load_base.(.type = NULL, .ed = NULL, .dir = NULL)
}
\arguments{
\item{.dir}{En el caso que se quiera cargar un archivo de SPSS de forma local se le debe de indicar la ruta.}

\item{.eaii}{Serie historica con el formato : AñoInicial-AñoFinal.}
}
\value{
Devuelve un tibble.
}
\description{
Crear un grupo para luego particionar la poblaci'on y calcular el total poblacional en cada momento
}
\examples{

load_base.("2016-2018")

load_base.(
  .dir = here::here(
    "data-raw",
    "2016-2018.sav"
  )
)
}
\keyword{utils}
