% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaii-recipes.R
\name{load_recipes_eaii.}
\alias{load_recipes_eaii.}
\title{Cargar recetas de la EAII}
\usage{
load_recipes_eaii.(.data, .eaii, .recipe = "all")
}
\arguments{
\item{.data}{Tibble con edición de la EAII.}

\item{.eaii}{Edición de la encuesta}

\item{.recipe}{Crear grupo de variables}
}
\value{
Devuelve un tibble.
}
\description{
Cargar recetas de la EAII
}
\examples{

load_base.("2016-2018")

load_base.(
  .dir = here::here(
    "data-raw",
    "2016-2018.sav"
  )
)
}
\keyword{internal}
