% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaii-recipes.R
\name{load_recipes_ech.}
\alias{load_recipes_ech.}
\title{Load and apply the recipe from a database table to an ECH}
\usage{
load_recipes_ech.(.data, .ech, .recipe, zap_labels = TRUE)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.recipe}{Recipe name}

\item{.ed}{Survey year}

\item{.zap_labels}{Boolean of labels to be removed}
}
\value{
Returns a tibble with the created variables
}
\description{
Load and apply the recipe from a database table to an ECH
}
\examples{



load_base.(
  .dir = here::here(
    "P_2019_Terceros.sav"
  )
) \%>\%
  load_recipes_ech.(
    .ech = "2019",
    .recipe = "labor"
  )

}
\keyword{utils}
