% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaii-recipes.R
\name{load_recipes_from_db.}
\alias{load_recipes_from_db.}
\title{Load and apply the recipe from a database table}
\usage{
load_recipes_from_db.(.data, .cnn, .survey, .ed, .recipe)
}
\arguments{
\item{.data}{Tibble with the survey}

\item{.cnn}{Database connection}

\item{.ed}{Survey year}

\item{.recipe}{Recipe name}
}
\value{
Returns a tibble with the created variables
}
\description{
Load and apply the recipe from a database table
}
\examples{



load_base.(
  .dir = here::here(
    "data-raw",
    "2016-2018.sav"
  )
) \%>\%
  load_recipes_from_db.(
    .cnn = example,
    .ed = 2016 - 2018,
    .recipe = "realiza_innovacion"
  )

}
\keyword{utils}
