% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaii-recipes.R
\name{new_recipe}
\alias{new_recipe}
\title{Load and apply the recipe from a database table}
\usage{
new_recipe(.type, .ed, .formula, ...)
}
\value{
Returns a tibble with the created variables
}
\description{
Load and apply the recipe from a database table
}
\examples{



load_base.(
  .dir = here::here(
    "data-raw",
    "2016-2018.sav"
  )
) \%>\%
  load_recipes_from_db.(
    .cnn = example,
    .ed = 2016 - 2018,
    .recipe = "realiza_innovacion"
  )

}
\keyword{utils}
