% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-normalize_names.R
\name{normalize_names.}
\alias{normalize_names.}
\title{Normalize names}
\usage{
normalize_names.(.data, meth = tolower)
}
\arguments{
\item{.data}{A tibble.}

\item{.meth}{Nombre de la variable que vamos a crear.}
}
\value{
Devuelve un tibble.
}
\description{
Normalize names
}
\keyword{utils}
