% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_weight.R
\name{set_weight.}
\alias{set_weight.}
\title{Set a weight}
\usage{
set_weight.(.data, .eaii = NULL, .ech = NULL, .weight = weight)
}
\arguments{
\item{.data}{A tibble.}

\item{.name}{Nombre de la variable que vamos a crear.}

\item{.keep}{keep variabels}
}
\value{
Devuelve un tibble.
}
\description{
Set a weight
}
\keyword{utils}
