% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy.eaii-package.R
\docType{package}
\name{srvyuRu-package}
\alias{srvyuRu}
\alias{srvyuRu-package}
\title{srvyuRu: Uruguay survey processing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mauro Loprete \email{mloprete@cognus.cl}

Authors:
\itemize{
  \item COGNUS [copyright holder, funder]
}

}
\keyword{internal}
