% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-estimate_totals.R
\name{stats_counts.}
\alias{stats_counts.}
\alias{stats_totals.}
\title{Estimate a count's from a sample of data with H-T}
\usage{
stats_counts.(.data, .variable = NULL, .by = NULL, .filter_labels = NULL)

stats_totals.(.data, .vars, .by = NULL, .filter_labels = NULL)
}
\arguments{
\item{.variable}{variable}

\item{.by}{a variable to agroup the data}

\item{.weight}{weight}

\item{.filer_labels}{labels to filter}
}
\description{
Estimate a count's from a sample of data with H-T

Estimate a total from a sample of data with H-T
}
\keyword{stats}
