% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-estimate_ratio.R
\name{stats_ratio.}
\alias{stats_ratio.}
\title{Estimate a ratio from a sample of data with H-T}
\usage{
stats_ratio.(.data, .num, .den, .name = ratio, .by = NULL, .pivot = TRUE)
}
\arguments{
\item{.num}{numerator}

\item{.den}{denominator}

\item{.name}{name of the ratio}

\item{.by}{a variable to group the data}

\item{.pivot}{}
}
\description{
Estimate a ratio from a sample of data with H-T
}
\keyword{stats}
