% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-to_oslo4.R
\name{to_oslo4.}
\alias{to_oslo4.}
\title{Crear un grupo para luego particionar la poblaci'on y
calcular el total poblacional en cada momento}
\usage{
to_oslo4.(.data)
}
\arguments{
\item{.data}{A tibble.}

\item{.name}{Nombre de la variable que vamos a crear.}

\item{...}{Condiciones para armar dicha variable}

\item{.keep}{}
}
\value{
Devuelve un tibble.
}
\description{
Crear un grupo para luego particionar la poblaci'on y
calcular el total poblacional en cada momento
}
\examples{
load_base.(
  "2016-2018"
) \%>\%
  recode(
    var1,
   IB_3.1.1 == 'pri' ~ 'Privada',
    TRUE ~ 'Publica/Mixta'
    .ncond = "',"
  ) \%>\%
  recode(
    var2,
   RAI_E.1_1 == 1 ~ 'Si',
   RAI_E.1_2 == 1 ~ 'Si',
   TRUE ~ 'No',
   .keep = 'none' # Con keep = 'none' nos quedamos solo con las varaibles que creamos
  ) \%>\%
  tidytable::select.(
    var1,
    var2
  )
}
\keyword{internal}
