% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-to_personas.R
\name{to_personas.}
\alias{to_personas.}
\title{Realizar estimaciones a nivel de persona para la ECH}
\usage{
to_personas.(.data, .ech)
}
\arguments{
\item{.data}{A tibble.}
}
\value{
Devuelve un tibble.
}
\description{
Realizar estimaciones a nivel de persona para la ECH
}
\examples{
load_base.(
  "2016-2018"
) \%>\%
  recode(
    var1,
   IB_3.1.1 == 'pri' ~ 'Privada',
    TRUE ~ 'Publica/Mixta'
    .ncond = "',"
  ) \%>\%
  recode(
    var2,
   RAI_E.1_1 == 1 ~ 'Si',
   RAI_E.1_2 == 1 ~ 'Si',
   TRUE ~ 'No',
   .keep = 'none' # Con keep = 'none' nos quedamos solo con las varaibles que creamos
  ) \%>\%
  tidytable::select.(
    var1,
    var2
  )
}
\keyword{ech}
\keyword{utils}
