% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validar_serie.R
\name{validar_serie.}
\alias{validar_serie.}
\title{Validar serie}
\usage{
validar_serie.(.svy, .vars)
}
\arguments{
\item{.vars}{variable.}

\item{.data}{A tibble.}
}
\value{
Devuelve un tibble.
}
\description{
Validar serie
}
\keyword{utils}
